Move ResourceStatus to dynamic-plugin-sdk
