+++ frontend/packages/console-dynamic-plugin-sdk/src/app/components/index.ts
@@ -1,0 +1,1 @@
+export { default as ResourceStatus } from './utils/resource-status';
