+++ frontend/packages/console-dynamic-plugin-sdk/src/app/index.ts
@@ -1,4 +1,5 @@
 export { default as AppInitSDK } from './AppInitSDK';
 export { SDKReducers } from './redux';
+export * from './components';
 export * from './core';
 export * from './redux-types';
