+++ frontend/packages/operator-lifecycle-manager/src/components/subscription.tsx
@@ -6,6 +6,7 @@ import * as classNames from 'classnames';
 import * as _ from 'lodash';
 import { useTranslation } from 'react-i18next';
 import { match, Link } from 'react-router-dom';
+import { ResourceStatus } from '@console/dynamic-plugin-sdk';
 import { Conditions } from '@console/internal/components/conditions';
 import {
   DetailsPage,
@@ -23,7 +24,6 @@ import {
   ResourceKebab,
   ResourceLink,
   resourcePathFromModel,
-  ResourceStatus,
   ResourceSummary,
   SectionHeading,
 } from '@console/internal/components/utils';
